# openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Marketplace AI API",
    "version": "1.0.0",
    "description": "API documentation for the Marketplace AI platform"
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Development server"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      },
      "apiKeyAuth": {
        "type": "apiKey",
        "in": "header",
        "name": "X-API-Key",
        "description": "API key for webhook authentication"
      }
    },
    "schemas": {
...
      "Error": {
        "type": "object",
        "properties": {
          "status": {
            "type": "string",
            "description": "Error status"
          },
          "message": {
            "type": "string",
            "description": "Error message"
          },
          "errors": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "field": {
                  "type": "string"
                },
                "message": {
                  "type": "string"
                }
              }
            },
            "description": "Validation errors"
          }
        }
      },
...
      "RentalRequest": {
        "type": "object",
        "required": [
          "startDate",
          "endDate"
        ],
        "properties": {
          "startDate": {
            "type": "string",
            "format": "date-time",
            "description": "Rental start date"
          },
          "endDate": {
            "type": "string",
            "format": "date-time",
            "description": "Rental end date"
          },
          "specialRequests": {
            "type": "string",
            "description": "Any special requests or requirements"
          }
        }
      },
...
      "CancelRequest": {
        "type": "object",
        "required": [
          "reason"
        ],
        "properties": {
          "reason": {
            "type": "string",
            "description": "Reason for cancellation"
          }
        }
      },
      "SuccessResponse": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string",
            "description": "Success message"
          },
          "data": {
            "type": "object",
            "description": "Response data"
          }
        }
      },
...
    }
  },
  "security": [
    {
      "bearerAuth": []
    }
  ],
  "paths": {
...
    "/api/rentals/{listingId}/request": {
      "post": {
        "summary": "Request a rental",
        "tags": [
          "Rentals"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "listingId",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Listing ID"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RentalRequest"
              },
              "examples": {
                "RequestRental": {
                  "$ref": "#/components/examples/RequestExamples/RequestRental"
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Rental request submitted",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessResponse"
                },
                "examples": {
                  "RentalRequested": {
                    "$ref": "#/components/examples/ResponseExamples/RentalRequested"
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid input or rental not allowed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                },
                "examples": {
                  "ValidationError": {
                    "$ref": "#/components/examples/ErrorExamples/ValidationError"
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                },
                "examples": {
                  "AuthenticationError": {
                    "$ref": "#/components/examples/ErrorExamples/AuthenticationError"
                  }
                }
              }
            }
          },
          "404": {
            "description": "Listing not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                },
                "examples": {
                  "NotFoundError": {
                    "$ref": "#/components/examples/ErrorExamples/NotFoundError"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/rentals/{requestId}/approve": {
      "post": {
        "summary": "Approve a rental request",
        "tags": [
          "Rentals"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "requestId",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Rental request ID"
          }
        ],
        "responses": {
          "200": {
            "description": "Rental request approved",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessResponse"
                },
                "examples": {
                  "SuccessResponse": {
                    "value": {
                      "message": "Rental request approved",
                      "data": null
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid request or rental not allowed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                },
                "examples": {
                  "InvalidStateError": {
                    "$ref": "#/components/examples/ErrorExamples/InvalidStateError"
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                },
                "examples": {
                  "AuthenticationError": {
                    "$ref": "#/components/examples/ErrorExamples/AuthenticationError"
                  }
                }
              }
            }
          },
          "404": {
            "description": "Rental request not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                },
                "examples": {
                  "NotFoundError": {
                    "$ref": "#/components/examples/ErrorExamples/NotFoundError"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/rentals/{requestId}/reject": {
      "post": {
        "summary": "Reject a rental request",
        "tags": [
          "Rentals"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "requestId",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Rental request ID"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CancelRequest"
              },
              "examples": {
                "CancelRequest": {
                  "value": {
                    "reason": "Renter did not meet requirements."
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Rental request rejected",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessResponse"
                },
                "examples": {
                  "SuccessResponse": {
                    "value": {
                      "message": "Rental request rejected",
                      "data": null
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid request or rental not allowed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                },
                "examples": {
                  "InvalidStateError": {
                    "$ref": "#/components/examples/ErrorExamples/InvalidStateError"
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                },
                "examples": {
                  "AuthenticationError": {
                    "$ref": "#/components/examples/ErrorExamples/AuthenticationError"
                  }
                }
              }
            }
          },
          "404": {
            "description": "Rental request not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                },
                "examples": {
                  "NotFoundError": {
                    "$ref": "#/components/examples/ErrorExamples/NotFoundError"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/rentals/{rentalId}/extend": {
      "post": {
        "summary": "Request a rental extension",
        "tags": [
          "Rentals"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "rentalId",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Rental ID"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ExtensionRequest"
              },
              "examples": {
                "ExtendRental": {
                  "$ref": "#/components/examples/RequestExamples/ExtendRental"
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Rental extension requested",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessResponse"
                },
                "examples": {
                  "RentalExtended": {
                    "$ref": "#/components/examples/ResponseExamples/RentalExtended"
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid request or extension not allowed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                },
                "examples": {
                  "InvalidStateError": {
                    "$ref": "#/components/examples/ErrorExamples/InvalidStateError"
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                },
                "examples": {
                  "AuthenticationError": {
                    "$ref": "#/components/examples/ErrorExamples/AuthenticationError"
                  }
                }
              }
            }
          },
          "404": {
            "description": "Rental not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                },
                "examples": {
                  "NotFoundError": {
                    "$ref": "#/components/examples/ErrorExamples/NotFoundError"
                  }
                }
              }
            }
          }
        }
      }
    },
...
  },
...
}